Moved remaining services to services directory
