+++ src/gluon/commands/jenkins/JenkinsBuild.ts
@@ -9,8 +9,8 @@ import {
 import * as _ from "lodash";
 import {QMConfig} from "../../../config/QMConfig";
 import {GluonService} from "../../services/gluon/GluonService";
-import {JenkinsService} from "../../util/jenkins/Jenkins";
-import {OCService} from "../../util/openshift/OCService";
+import {JenkinsService} from "../../services/jenkins/JenkinsService";
+import {OCService} from "../../services/openshift/OCService";
 import {menuForApplications} from "../../util/packages/Applications";
 import {menuForProjects} from "../../util/project/Project";
 import {
