+++ src/gluon/commands/project/AddConfigServer.ts
@@ -11,7 +11,7 @@ import {SlackMessage, url} from "@atomist/slack-messages";
 import * as _ from "lodash";
 import {QMConfig} from "../../../config/QMConfig";
 import {GluonService} from "../../services/gluon/GluonService";
-import {OCService} from "../../util/openshift/OCService";
+import {OCService} from "../../services/openshift/OCService";
 import {handleQMError, ResponderMessageClient} from "../../util/shared/Error";
 import {
     RecursiveParameter,
